Refactor : Option, OptionGroup 전체 패키지 경로 조정
- Option / OptionGroup => options 패키지 하위에서 option, optionGroup으로 분류되도록 패키지 경로 조정
