+++ src/main/java/team/mosk/api/server/domain/options/option/error/OptionNotFoundException.java
@@ -1,4 +1,4 @@
-package team.mosk.api.server.domain.option.error;
+package team.mosk.api.server.domain.options.option.error;
 
 public class OptionNotFoundException extends RuntimeException{
     public OptionNotFoundException(String message) {
