+++ src/main/java/team/mosk/api/server/domain/options/optionGroup/dto/OptionGroupResponse.java
@@ -1,8 +1,8 @@
-package team.mosk.api.server.domain.option.dto;
+package team.mosk.api.server.domain.options.optionGroup.dto;
 
 import lombok.*;
-import team.mosk.api.server.domain.option.model.persist.Option;
-import team.mosk.api.server.domain.option.model.persist.OptionGroup;
+import team.mosk.api.server.domain.options.option.model.persist.Option;
+import team.mosk.api.server.domain.options.optionGroup.model.persist.OptionGroup;
 import team.mosk.api.server.domain.product.model.persist.Product;
 
 import java.util.List;
