+++ src/main/java/team/mosk/api/server/domain/options/optionGroup/error/OptionGroupNotFoundException.java
@@ -1,4 +1,4 @@
-package team.mosk.api.server.domain.option.error;
+package team.mosk.api.server.domain.options.optionGroup.error;
 
 public class OptionGroupNotFoundException extends RuntimeException{
     public OptionGroupNotFoundException(String message) {
