+++ src/main/java/team/mosk/api/server/domain/options/optionGroup/model/persist/OptionGroup.java
@@ -1,7 +1,8 @@
-package team.mosk.api.server.domain.option.model.persist;
+package team.mosk.api.server.domain.options.optionGroup.model.persist;
 
 import lombok.*;
-import team.mosk.api.server.domain.option.dto.UpdateOptionGroupRequest;
+import team.mosk.api.server.domain.options.option.model.persist.Option;
+import team.mosk.api.server.domain.options.optionGroup.dto.UpdateOptionGroupRequest;
 import team.mosk.api.server.domain.product.model.persist.Product;
 import team.mosk.api.server.global.common.BaseEntity;
