+++ src/main/java/team/mosk/api/server/domain/options/optionGroup/model/persist/OptionGroupRepository.java
@@ -1,4 +1,4 @@
-package team.mosk.api.server.domain.option.model.persist;
+package team.mosk.api.server.domain.options.optionGroup.model.persist;
 
 import org.springframework.data.jpa.repository.JpaRepository;
